moved database queries into data folder
